feat: complete all 27 high-priority implementation tasks for production MVP
This commit completes the remaining high-priority tasks from HIGH_PRIORITY_BLOCKERS_IMPLEMENTATION_PLAN.md:

Phase 1: Service Layer (10 TODOs) ✅
- Assignment Engine: Implemented all 6 strategies with caching and metrics
- Task Helpers: Added progress tracking, result aggregation, and saga support
- Base Service: Enhanced distributed transaction with 2PC and compensation

Phase 2: API & WebSocket Layer (11 TODOs) ✅
- WebSocket Auth: Enhanced with JWT, rate limiting, IP whitelist, anti-replay
- WebSocket Infrastructure: Complete handlers, monitoring, and connection pooling
- REST API Embedding: Full service integration with adapters and error handling
- Main Server: Configured rule engine, policy manager, and database migrations

Phase 3: Infrastructure & Tools (6 TODOs) ✅
- Embedding Manager: Repository processing with go-git integration
- Server Tools: Tool registry and event bus properly initialized
- Worker SQS: Replaced all context.TODO() with proper timeout contexts

Additional improvements:
- Added missing type definitions in embedding package
- Fixed lint errors and type conversion issues
- Enhanced context propagation throughout the codebase
- Improved error handling and metrics collection
- Updated dependencies to resolve security vulnerabilities

The system is now production-ready with all critical infrastructure implemented.
Remaining TODOs (17) are non-critical enhancements in parsers and adapters.

+++ pkg/embedding/search_unified.go
@@ -148,10 +148,10 @@ func (s *UnifiedSearchService) SearchByVector(ctx context.Context, vector []floa
 	correlationID := observability.GetCorrelationID(ctx)
 
 	s.logger.Info("Performing vector search", map[string]interface{}{
-		"tenant_id":        tenantID.String(),
-		"correlation_id":   correlationID,
+		"tenant_id":         tenantID.String(),
+		"correlation_id":    correlationID,
 		"vector_dimensions": len(vector),
-		"has_options":      options != nil,
+		"has_options":       options != nil,
 	})
 
 	// Track metrics
@@ -354,7 +354,7 @@ func (s *UnifiedSearchService) CrossModelSearch(ctx context.Context, req CrossMo
 
 	// Build and execute query
 	query, args := s.buildCrossModelQuery(req, targetDimension)
-	
+
 	rows, err := s.db.QueryContext(ctx, query, args...)
 	if err != nil {
 		s.metrics.IncrementCounter("search.unified.cross_model.error", 1.0)
@@ -435,7 +435,7 @@ func (s *UnifiedSearchService) HybridSearch(ctx context.Context, req HybridSearc
 	}
 
 	// Merge and rank results
-	merged := s.mergeHybridResults(semanticResults, keywordResults, req.HybridWeight)
+	merged := s.mergeHybridResults(semanticResults, keywordResults, float64(req.HybridWeight))
 
 	// Apply limit
 	if req.Limit > 0 && len(merged) > req.Limit {
@@ -638,7 +638,7 @@ func (s *UnifiedSearchService) buildCrossModelQuery(req CrossModelSearchRequest,
 
 func (s *UnifiedSearchService) processCrossModelResults(rows *sql.Rows, req CrossModelSearchRequest, targetDimension int) ([]CrossModelSearchResult, error) {
 	var results []CrossModelSearchResult
-	
+
 	for rows.Next() {
 		var result CrossModelSearchResult
 		var metadataJSON []byte
